+++ src/main/java/ezcli/modules/ezcli_core/Module.java
@@ -1,29 +1,46 @@
 package ezcli.modules.ezcli_core;
 
+import ezcli.modules.ezcli_core.terminal.Terminal;
+
+import java.io.IOException;
 import java.lang.reflect.Method;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.LinkedList;
+import java.util.List;
 
 /**
- * Abstract class specifying methods all modules should contain.
+ * Abstract class specifying methods all modules should contain, and some initialization ones.
  * Each module must extend this class and implement the methods specified here.
+ *
+ * <p>Two types of module exist:
+ *
+ * <p>One type is like the Terminal module, which is an independent entity
+ * that handles its own input and does not rely on other modules to function or for data of any type.
+ * That type of module should be initialized with the single parameter super class constructor, and the
+ * single parameter super class init method (in that order).
+ *
+ * <p>The other type of module that exists works on top of independent modules. It provides some extra
+ * optional functionality, which is non essential and can be removed if desired. These modules are run
+ * when certain events are detected. This type of module should be initialized by using the double parameter
+ * super class constructor and the triple parameter super class init method (in that order).
  */
 public abstract class Module {
 
-    // List of modules that the program will use
+    // Hashmap that maps modules to their class name (the class that extends Module)
     public static HashMap<String, Module> modules = new HashMap<>();
 
     // Hashmap that maps a string to a module, which when typed in interactive module, will run the mapped module
-    protected static HashMap<String, Module> moduleMap = new HashMap<>();
-    protected static HashMap<Module, String> moduleNameMap = new HashMap<>();
+    public static HashMap<String, Module> moduleMap = new HashMap<>();
+    private static HashMap<Module, String> moduleNameMap = new HashMap<>();
 
-    protected static HashMap<String, LinkedList<Method>> eventMethods = new HashMap<>();
+    // Contains methods to be run when events from TermKeyProcessor and TermArrowKeyProcessor are processed
+    private static HashMap<String, LinkedList<Method>> eventMethods = new HashMap<>();
 
     // Module specific variables...
-
     public String moduleName;
-
-    protected EventState whenToRun;
+    private EventState whenToRun;
 
     public Module(String moduleName) {
         this.moduleName = moduleName;
@@ -34,13 +51,9 @@ public abstract class Module {
         this.whenToRun = whenToRun;
     }
 
-    /**
-     * Return the preferred time to run for this module, so that the given
-     * method uses the correct and up to date variables.
-     *
-     * @return Preferred time to run module.
-     */
-    public abstract EventState getWhenToRun();
+    public EventState getWhenToRun() {
+        return whenToRun;
+    }
 
     /**
      * Code to run for module. Should contain a while loop similar to that in the Interactive class,
@@ -52,6 +65,39 @@ public abstract class Module {
      * Code to run when touring program.
      */
     public abstract void tour();
+
+    /**
+     * Initializes terminal module and all modules listed in moduleDeclaration file sequentially.
+     * For comments or temporarily disabling a module, put a '#' at the beginning of the module declaration line.
+     *
+     * @param moduleDeclarations File containing list of modules to be loaded
+     */
+    public static void initModules(String moduleDeclarations) {
+
+        new Terminal();
+
+        List<String> modules;
+
+        try {
+            modules = Files.readAllLines(Paths.get(moduleDeclarations));
+        } catch (IOException e) {
+            System.err.println("Modules could not be loaded, file \"" + moduleDeclarations + "\" not found");
+            return;
+        }
+
+        for (String s : modules) {
+            if ("".equals(s.trim()) || s.contains("#"))
+                continue;
+
+            try {
+                Class<?> module = Class.forName("ezcli.modules." + s.trim());
+                module.getConstructor().newInstance();
+            } catch (Exception e) {
+                System.err.println("Error loading module: ezcli.modules." + s.trim());
+                e.printStackTrace();
+            }
+        }
+    }
 
     /**
      * Quasi-constructor. Initializes the class and adds the module to a list of modules.
@@ -70,50 +116,100 @@ public abstract class Module {
      * Quasi-constructor. Adds methods to eventMethods, which are called when certain events occur in Terminal module,
      * such as key press events.
      */
-    protected void init(Module module, Method[] methods, String[] binds) {
+    protected void init(Module module, String[] methodNames, String[] binds) {
 
-        if (methods.length != binds.length) {
-            System.err.println("Methods and Binds arrays are not equal in length " +
-                    "(each method must be bound to something)");
+        if (methodNames.length != binds.length) {
+            System.err.println("Methods and Binds arrays have unequal size in module \""
+                    + module.getClass().getSimpleName() + "\" (each method must be bound to something) ");
             return;
+        }
+
+        Method[] methods = new Method[methodNames.length];
+        for (int i = 0; i < methodNames.length; i++) {
+            try {
+                methods[i] = module.getClass().getDeclaredMethod(methodNames[i]);
+            } catch (NoSuchMethodException e) {
+                System.err.println("Error loading method \"" + methodNames[i] + "\" from module "
+                        + module.getClass().getSimpleName());
+            }
         }
 
         modules.put(module.moduleName, module);
 
+        // Match key binds with methods passed in arrays
         for (int i = 0; i < methods.length; i++) {
+            if (methods[i] == null)
+                continue;
+
             String[] bind = binds[i].split(" ");
             for (String b : bind) {
-                if ("allkeys".equals(b)) {
-                    for (int j = 30; j < 126; j++)
-                        addCharsToHashmap(String.valueOf((char) j), methods[i]);
+                b = b.trim();
+                if (!bindToKeys(b, binds[i], methods[i]))
+                    bindToArrowKeys(b, methods[i]);
 
-                    String[] keys = {"\n", "\b", "\t"};
-                    for (String s : keys)
-                        addCharsToHashmap(s, methods[i]);
-
-                } else if (!b.contains("arrow")) {
-                    addCharsToHashmap(binds[i], methods[i]);
-                    continue;
-                }
-
-                if ("allarrows".equals(b)) {
-                    String[] arrows = {"uarrow", "darrow", "larrow", "rarrow"};
-                    for (String s : arrows)
-                        addStringToHashmap(s, methods[i]);
-                } else {
-                    if ("uarrow".equals(b))
-                        addStringToHashmap(b, methods[i]);
-                    if ("darrow".equals(b))
-                        addStringToHashmap(b, methods[i]);
-                    if ("larrow".equals(b))
-                        addStringToHashmap(b, methods[i]);
-                    if ("rarrow".equals(b))
-                        addStringToHashmap(b, methods[i]);
-                }
             }
         }
     }
 
+    /**
+     * Binds a sequence of characters to a method. If any character in the sequence
+     * if processed later on, the passed method will be invoked.
+     *
+     * @param b Current request to bind
+     * @param bind Potential sequence of characters to bind
+     * @param m Method to be bound to sequence of characters
+     * @return True if binding was attempted, false if no binding occurred
+     */
+    private boolean bindToKeys(String b, String bind, Method m) {
+        if ("allkeys".equals(b)) {
+            for (int j = 30; j < 126; j++) // ascii
+                addCharsToHashmap(String.valueOf((char) j), m);
+
+            String[] keys = {"\n", "\b", "\t"};
+            for (String s : keys)
+                addCharsToHashmap(s, m);
+            return true;
+
+        } else if (!b.contains("arrow")) {
+            addCharsToHashmap(bind, m);
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Binds a string to a method. If the string is passed as an event later on,
+     * the passed method will be invoked. For use primarily with keys that have no ASCII values.
+     *
+     * @param b String to bind to method
+     * @param m Method to be bound to string
+     */
+    private void bindToArrowKeys(String b, Method m) {
+        if ("allarrows".equals(b)) {
+            String[] arrows = {"uarrow", "darrow", "larrow", "rarrow"};
+            for (String s : arrows)
+                addStringToHashmap(s, m);
+        } else {
+            if ("uarrow".equals(b))
+                addStringToHashmap(b, m);
+            if ("darrow".equals(b))
+                addStringToHashmap(b, m);
+            if ("larrow".equals(b))
+                addStringToHashmap(b, m);
+            if ("rarrow".equals(b))
+                addStringToHashmap(b, m);
+        }
+    }
+
+    /**
+     * Breaks string into individual characters, and assigns each one the passed method, for later invocation when
+     * the relevant characters are handled.
+     * <p>
+     * Only for use with ASCII characters, non ASCII characters should be added with addStringToHashmap()
+     *
+     * @param s String of characters to assign to method
+     * @param m Method to be assigned to sequence of characters
+     */
     private void addCharsToHashmap(String s, Method m) {
         for (int i = 0; i < s.length(); i++) {
             LinkedList<Method> list = eventMethods.get(String.valueOf(s.charAt(i)));
@@ -127,6 +223,18 @@ public abstract class Module {
         }
     }
 
+    /**
+     * Assigns string to method, for later invocation when the string is handled in processEvent()
+     * <p>
+     * Only for use of keys which do not have ASCII codes, such as arrow keys.
+     * <p>
+     * Currently only supports uarrow, darrow, larrow and rarrow for arrow keys.
+     *
+     * @see ezcli.modules.ezcli_core.terminal.TermArrowKeyProcessor
+     *
+     * @param s String to map to method.
+     * @param m Method to be mapped to string
+     */
     private void addStringToHashmap(String s, Method m) {
         LinkedList<Method> list = eventMethods.get(s);
 
@@ -138,22 +246,33 @@ public abstract class Module {
         list.add(m);
     }
 
-    public static void processEvent(String val, EventState es) {
-        LinkedList<Method> list = eventMethods.get(val);
+    /**
+     * Searches eventMethods hashmap for list of modules to run based on event, and runs them.
+     * For more info regarding the calling of these methods and the hardcoded events that are accepted,
+     * see the mentioned classes.
+     *
+     * @see ezcli.modules.ezcli_core.terminal.TermKeyProcessor
+     * @see ezcli.modules.ezcli_core.terminal.TermArrowKeyProcessor
+     *
+     * @param event Event to process
+     * @param es Required stage to process at (currently either before or after the input processors have updated)
+     */
+    public static void processEvent(String event, EventState es) {
+        LinkedList<Method> list = eventMethods.get(event);
 
         if (list == null)
             return;
 
         for (Method m : list) {
             try {
-                System.out.println(m.getDeclaringClass().getSimpleName());
-                EventState requestedES = (EventState) m.getDeclaringClass().getDeclaredMethod("getWhenToRun")
-                        .invoke(modules.get(m.getDeclaringClass().getSimpleName()));
-                if (es == requestedES)
+                EventState requestedES =
+                        (EventState) m.getDeclaringClass().getSuperclass().getDeclaredMethod("getWhenToRun")
+                                .invoke(modules.get(m.getDeclaringClass().getSimpleName()));
+                if (es.equals(requestedES))
                     m.invoke(modules.get(m.getDeclaringClass().getSimpleName()));
             } catch (Exception e) {
                 System.err.println("Error processing event from class: " + m.getDeclaringClass().getSimpleName());
-                System.out.println("Error processing \"" + val + "\"");
+                System.out.println("Error processing \"" + event + "\"");
                 e.printStackTrace();
             }
         }